chore: endpoint 수정
