+++ src/main/java/com/be3c/sysmetic/domain/strategy/controller/StrategyController.java
@@ -1,9 +1,6 @@
 package com.be3c.sysmetic.domain.strategy.controller;
 
-import com.be3c.sysmetic.domain.strategy.dto.AccountImageRequestDto;
 import com.be3c.sysmetic.domain.strategy.dto.AccountImageResponseDto;
-import com.be3c.sysmetic.domain.strategy.service.*;
-import com.be3c.sysmetic.global.common.response.ErrorCode;
 import com.be3c.sysmetic.domain.strategy.service.*;
 import com.be3c.sysmetic.domain.strategy.dto.*;
 import com.be3c.sysmetic.global.common.response.PageResponse;
@@ -13,19 +10,16 @@ import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.persistence.EntityNotFoundException;
-import jakarta.validation.constraints.NotBlank;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDate;
-import java.util.List;
 
 @Tag(name = "전략 API", description = "사용자 전략")
+@RequestMapping("/v1/strategy")
 @RequiredArgsConstructor(onConstructor_ = @__(@Autowired))
 @RestController
 public class StrategyController {
@@ -41,7 +35,7 @@ public class StrategyController {
             summary = "매매방식, 종목 조회",
             description = "사용중인 모든 매매방식 및 종목 조회"
     )
-    @GetMapping("/strategy/method-and-stock")
+    @GetMapping("/method-and-stock")
     public ResponseEntity<APIResponse<MethodAndStockGetResponseDto>> findMethodAndStockList() {
         return ResponseEntity.status(HttpStatus.OK)
                 .body(APIResponse.success(strategyService.findMethodAndStock()));
@@ -52,12 +46,12 @@ public class StrategyController {
             summary = "일간분석 조회",
             description = "일간분석 목록 조회"
     )
-    @GetMapping("/strategy/daily")
+    @GetMapping("/daily")
     public ResponseEntity<APIResponse<PageResponse<DailyGetResponseDto>>> findDaily(
             @RequestParam("strategyId") Long strategyId,
             @RequestParam("page") Integer page,
-            @RequestParam(value = "startDate", required = false) LocalDate startDate,
-            @RequestParam(value = "endDate", required = false) LocalDate endDate
+            @RequestParam(value = "startDate", required = false) @Schema(description = "조회 시작 년월일", example = "2024-11-01") LocalDate startDate,
+            @RequestParam(value = "endDate", required = false) @Schema(description = "조회 종료 년월일", example = "2024-11-26") LocalDate endDate
     ) {
         return ResponseEntity.status(HttpStatus.OK)
                 .body(APIResponse.success(dailyService.findDaily(strategyId, page, startDate, endDate)));
@@ -68,7 +62,7 @@ public class StrategyController {
             summary = "월간분석 조회",
             description = "월간분석 목록 조회"
     )
-    @GetMapping("/strategy/monthly")
+    @GetMapping("/monthly")
     public ResponseEntity<APIResponse<PageResponse<MonthlyGetResponseDto>>> findMonthly(
             @RequestParam("strategyId") Long strategyId,
             @RequestParam("page") Integer page,
@@ -84,7 +78,7 @@ public class StrategyController {
             summary = "실계좌이미지 조회",
             description = "실계좌이미지 목록 조회"
     )
-    @GetMapping("/strategy/account-image")
+    @GetMapping("/account-image")
     public ResponseEntity<APIResponse<PageResponse<AccountImageResponseDto>>> getAccountImage(
             @RequestParam Long strategyId,
             @RequestParam Integer page
@@ -93,62 +87,15 @@ public class StrategyController {
         return ResponseEntity.status(HttpStatus.OK).body(APIResponse.success(responseDto));
     }
 
-    // 실계좌이미지 삭제
-    @Operation(
-            summary = "실계좌이미지 삭제",
-            description = "트레이더가 본인의 실계좌이미지 삭제"
-    )
-    @DeleteMapping("/strategy/account-image/{accountImageId}")
-    public ResponseEntity<APIResponse> deleteAccountImage(@PathVariable Long accountImageId) {
-        accountImageService.deleteAccountImage(accountImageId);
-        return ResponseEntity.ok(APIResponse.success());
-    }
-
-    // todo. 파일 관련 작업 필요. 예슬님이 이어서 작업해주실 예정입니다.
-    // 실계좌이미지 등록
-    @Operation(
-            summary = "실계좌이미지 등록",
-            description = "트레이더가 본인의 실계좌이미지 등록"
-    )
-    @PostMapping(value = "/strategy/account-image", consumes = {MediaType.MULTIPART_FORM_DATA_VALUE})
-    public ResponseEntity<APIResponse> saveAccountImage(
-            @RequestPart List<AccountImageRequestDto> accountImages,
-            @RequestParam Long strategyId
-    ) {
-        accountImageService.saveAccountImage(strategyId, accountImages);
-        return ResponseEntity.ok(APIResponse.success());
-    }
-
     // 통계 조회
     @Operation(
             summary = "통계 조회",
-            description = "특정 전략의 통계 조회",
-            responses = {
-                    @ApiResponse(content = @Content(schema = @Schema(implementation = StrategyStatisticsGetResponseDto.class)))
-            }
+            description = "특정 전략의 통계 조회"
     )
-    @GetMapping("/strategy/statistics/{strategyId}")
+    @GetMapping("/statistics/{strategyId}")
     public ResponseEntity<APIResponse<StrategyStatisticsGetResponseDto>> findStatistics(@PathVariable Long strategyId) {
         return ResponseEntity.status(HttpStatus.OK)
                 .body(APIResponse.success(strategyStatisticsService.findStrategyStatistics(strategyId)));
-    }
-
-    // 비공개 전환
-    @PatchMapping("/strategy/{id}/visibilaty")
-    public ResponseEntity<APIResponse<String>> patchStrategy(
-            @NotBlank @PathVariable Long id
-    ) {
-        try {
-            if(strategyService.privateStrategy(id)) {
-                return ResponseEntity.status(HttpStatus.OK)
-                        .body(APIResponse.success());
-            }
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(APIResponse.fail(ErrorCode.INTERNAL_SERVER_ERROR));
-        } catch (EntityNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body(APIResponse.fail(ErrorCode.NOT_FOUND));
-        }
     }
 
 }
